fix: move refetch interval dropdown
