+++ frontend/app/src/components/refetch-interval-dropdown.tsx
